# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/find/application/common/ApplicationFinder.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/after/DetailAfterRemandImpl.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.infra/src/main/java/nts/uk/ctx/at/request/infra/repository/application/common/JpaApplicationRepository_New.java
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/000/b/kaf000.b.start.ts

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/add/AddSpLea7InfoCommandHandler.java
@@ -14,12 +14,13 @@ import nts.uk.shr.pereg.app.command.PeregAddCommandHandler;
 import nts.uk.shr.pereg.app.command.PeregAddCommandResult;
 
 @Stateless
-public class AddSpLea7InfoCommandHandler extends CommandHandlerWithResult<AddSpecialleave7informationCommand, PeregAddCommandResult>
-implements PeregAddCommandHandler<AddSpecialleave7informationCommand>{
+public class AddSpLea7InfoCommandHandler
+		extends CommandHandlerWithResult<AddSpecialleave7informationCommand, PeregAddCommandResult>
+		implements PeregAddCommandHandler<AddSpecialleave7informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler addSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00031";
@@ -32,12 +33,13 @@ implements PeregAddCommandHandler<AddSpecialleave7informationCommand>{
 
 	@Override
 	protected PeregAddCommandResult handle(CommandHandlerContext<AddSpecialleave7informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00031.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
+
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00031.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
 				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		return new PeregAddCommandResult(addSpLeaInfoCommandHandler.addHandler(domain));
 	}
