+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpLea20InfoCommandHandler.java
@@ -14,11 +14,11 @@ import nts.uk.shr.pereg.app.command.PeregUpdateCommandHandler;
 
 @Stateless
 public class UpdateSpLea20InfoCommandHandler extends CommandHandler<UpdateSpecialleave20informationCommand>
-implements PeregUpdateCommandHandler<UpdateSpecialleave20informationCommand>{
+		implements PeregUpdateCommandHandler<UpdateSpecialleave20informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler updateSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00058";
@@ -31,12 +31,13 @@ implements PeregUpdateCommandHandler<UpdateSpecialleave20informationCommand>{
 
 	@Override
 	protected void handle(CommandHandlerContext<UpdateSpecialleave20informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid,command.getSID(),SpecialLeaveCode.CS00058.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
-				command.getGrantDays()!=null? command.getGrantDays().intValue() : null, command.getGrantTable());
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00058.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
+				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		updateSpLeaInfoCommandHandler.updateHandler(domain);
 	}
