+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/divergence/time/DivergenceTimeDto.java
@@ -101,7 +101,7 @@ public class DivergenceTimeDto implements DivergenceTimeSetMemento {
 	 * @see nts.uk.ctx.at.record.dom.divergence.time.DivergenceTimeSetMemento#setTarsetItems(java.util.List)
 	 */
 	@Override
-	public void setTarsetItems(List<Double> targetItems) {
+	public void setTarsetItems(List<Integer> targetItems) {
 		// no code
 
 	}
