+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/remainingnumber/specialleavegrant/SpecialLeaveGrantDto.java
@@ -20,12 +20,12 @@ public class SpecialLeaveGrantDto{
 	private GeneralDate deadlineDate;
 	private int expStatus;
 	private int registerType;
-	private int numberDayGrant;
+	private double numberDayGrant;
 	private int timeGrant;
 	private double numberDayUse;
 	private int timeUse;
 	private double useSavingDays;
-	private int numberDaysOver;
+	private double numberDaysOver;
 	private int timeOver;
 	private double numberDayRemain;
 	private int timeRemain;
